refactor(server): split test configuration fixtures
